+++ modules/github-repo/main.tf
@@ -116,3 +116,31 @@ resource "github_repository_pages" "this" {
     }
   }
 }
+
+# Legacy Branch Protections
+resource "github_branch_protection" "this" {
+  for_each = { for bp in coalesce(var.config.branch_protections, []) : trimspace(bp.branch) => bp }
+
+  repository_id                   = github_repository.this.node_id
+  pattern                         = each.key
+  enforce_admins                  = each.value.enforceAdmins
+  require_signed_commits          = each.value.requireSignedCommits
+  require_conversation_resolution = each.value.requireConversationResolution
+
+  dynamic "required_status_checks" {
+    for_each = length(coalesce(each.value.statusChecks, [])) > 0 ? [each.value.statusChecks] : []
+    content {
+      contexts = required_status_checks.value
+    }
+  }
+
+  dynamic "required_pull_request_reviews" {
+    for_each = (each.value.requiredReviewersCount > 0 || each.value.requiredCodeownersReviewers) ? [1] : []
+    content {
+      required_approving_review_count = each.value.requiredReviewersCount
+      require_code_owner_reviews      = each.value.requiredCodeownersReviewers
+    }
+  }
+
+  depends_on = [github_branch_default.this]
+}